+++ src/rer/sitesearch/browser/static/js/IndexesFilters/index.js
@@ -1,64 +1,67 @@
 import React, { useContext } from 'react';
-import Select, { components } from 'react-select';
 import SearchContext from '../utils/searchContext';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { icons } from '../utils/icons';
+import SelectField from './select';
+import BooleanField from './boolean';
+import DateField from './date';
 
 // faFolder
 const { faTags, faListUl } = icons;
 
 const IndexesFilters = () => {
   const { setFilters, filters, facets, results } = useContext(SearchContext);
-  const canShow = results.length > 0 && facets && facets.indexes ? true : false;
+  const facetsIndexes = facets ? facets.indexes : null;
+  const hasDateIndex = facetsIndexes
+    ? Object.values(facetsIndexes.values).filter(
+        index => index.type === 'DateIndex',
+      ).length > 0
+    : false;
+  const canShow =
+    hasDateIndex || (results.length > 0 && facetsIndexes) ? true : false;
   if (!canShow) {
     return '';
   }
 
-  const filtersData = facets.indexes.order.map((index, idx) => {
-    const facet = facets.indexes.values[index];
+  const filtersData = facetsIndexes.order.map((index, idx) => {
+    const facet = facetsIndexes.values[index];
     const facetValues = facet.values;
-    const options = Object.keys(facetValues).map(key => ({
-      value: key,
-      label: `${key} (${facetValues[key]})`,
-    }));
-    return Object.keys(facetValues).length > 0 ? (
+    let field = '';
+    switch (facet.type) {
+      case 'BooleanIndex':
+        field = (
+          <BooleanField
+            values={facetValues}
+            filters={filters}
+            index={index}
+            setFilters={setFilters}
+          />
+        );
+        break;
+      case 'DateIndex':
+        field = (
+          <DateField filters={filters} index={index} setFilters={setFilters} />
+        );
+        break;
+      default:
+        field = (
+          <SelectField
+            values={facetValues}
+            filters={filters}
+            index={index}
+            setFilters={setFilters}
+          />
+        );
+        break;
+    }
+    return facet.type === 'DateIndex' || Object.keys(facetValues).length > 0 ? (
       <div className="filter-item" key={index + idx}>
         <h3>
           {index === 'Subject' && <FontAwesomeIcon icon={faTags} />}
           {index === 'Temi' && <FontAwesomeIcon icon={faListUl} />}{' '}
           {facet.label}
         </h3>
-        <Select
-          options={options}
-          isMulti
-          isClearable
-          components={{
-            // eslint-disable-next-line react/display-name
-            MultiValueLabel: props => (
-              <components.MultiValueLabel {...props} className="text-primary" />
-            ),
-          }}
-          className="rer-sitesearch-select text-primary"
-          placeholder="Cerca per categorie"
-          aria-controls="sitesearch-results-list"
-          value={options.filter(option =>
-            filters[index]
-              ? filters[index].query.includes(option.value)
-              : false,
-          )}
-          onChange={option => {
-            if (!option || option.length == 0) {
-              setFilters({ [index]: '' });
-            } else {
-              setFilters({
-                [index]: {
-                  query: option.map(({ value }) => value),
-                  operator: 'and',
-                },
-              });
-            }
-          }}
-        />
+        {field}
       </div>
     ) : null;
   });