+++ src/rer/sitesearch/upgrades.py
@@ -8,3 +8,8 @@ def to_4000(context):
     """ """
     context.runAllImportStepsFromProfile("profile-rer.sitesearch:to_4000")
     context.runImportStepFromProfile(default_profile, "plone.app.registry")
+
+
+def to_4001(context):
+    """ """
+    context.runImportStepFromProfile(default_profile, "plone.app.registry")